throw exception in case entry does not exist

+++ app/Service/Population.php
@@ -72,6 +72,10 @@ class Population
     {
         $population = $this->get($id);
 
+        if (!$population) {
+            throw new StatusCode\NotFoundException('Entry does not exist');
+        }
+
         $this->populationTable->update([
             'id'          => $population['id'],
             'place'       => $place,
@@ -85,6 +89,10 @@ class Population
     public function delete($id)
     {
         $population = $this->get($id);
+
+        if (!$population) {
+            throw new StatusCode\NotFoundException('Entry does not exist');
+        }
 
         $this->populationTable->delete([
             'id' => $population['id']
